Ok, 5am, it works.  For real.

+++ mail_server/disable-ipv6.sh
@@ -3,8 +3,12 @@
 #if ip a | grep inet6; then
 if [ `cat /proc/sys/net/ipv6/conf/all/disable_ipv6` -ne 1 ]; then
     echo "Ipv6 detected, disabling"
-    file_path="/etc/sysctl.d/ipv6-disabled"
-    rm -f $file_path
+    if [ ! -f "/etc/sysctl.conf" ]; then
+        file_path="/etc/sysctl.conf"
+    else
+        file_path="/etc/sysctl.d/ipv6-disabled"
+        rm -f $file_path
+    fi
     echo "net.ipv6.conf.all.disable_ipv6 = 1" >> $file_path
     echo "net.ipv6.conf.default.disable_ipv6 = 1" >> $file_path
     echo "net.ipv6.conf.lo.disable_ipv6 = 1" >> $file_path
